refactor: move interface to folder
