+++ Web.Api/Controllers/UsuarioController.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Model.DTO;
-using Service;
+using Service.Interfaces;
 
 namespace Web.Api.Controllers
 {
